refactor(useEnvironment): hooks/client/配下へ移す
useEnvironmentはモジュールスコープでcreateContextを呼ぶためclient境界が必須。
EnvironmentProviderを公開しており利用者のServer Componentから直接
レンダリングされうるため、そのファイル自身に'use client'を持つ必要がある。

client境界が必要であることがディレクトリから読み取れるよう、
hooks/client/配下へ移動する。
