+++ packages/smarthr-ui/src/components/InputFile/FilePreviewDialog.tsx
@@ -2,7 +2,7 @@
 
 import { type FC, memo, useEffect, useState } from 'react'
 
-import { useEnvironment } from '../../hooks/useEnvironment'
+import { useEnvironment } from '../../hooks/client/useEnvironment'
 import { Localizer } from '../../intl'
 import { Button } from '../Button'
 import { Dialog, ModelessDialog } from '../Dialog'
