+++ packages/smarthr-ui/src/hooks/client/useEnvironment/index.ts
