+++ packages/smarthr-ui/src/hooks/client/useEnvironment/useEnvironment.stories.tsx
@@ -1,11 +1,11 @@
-import { DefinitionList, DefinitionListItem } from '../../components/DefinitionList'
-import { Heading } from '../../components/Heading'
-import { Stack } from '../../components/Layout'
-import { Groupbox } from '../../components/Panel'
-import { createTheme, defaultMediaQuery } from '../../themes'
-import { ThemeProvider } from '../useTheme'
+import { DefinitionList, DefinitionListItem } from '../../../components/DefinitionList'
+import { Heading } from '../../../components/Heading'
+import { Stack } from '../../../components/Layout'
+import { Groupbox } from '../../../components/Panel'
+import { createTheme, defaultMediaQuery } from '../../../themes'
+import { ThemeProvider } from '../../useTheme'
+import { EnvironmentProvider } from '../EnvironmentProvider'
 
-import { EnvironmentProvider } from './EnvironmentProvider'
 import { useEnvironment } from './useEnvironment'
 
 import type { Meta, StoryObj } from '@storybook/react/*'
