+++ packages/smarthr-ui/src/hooks/useSV/useSV.ts
@@ -1,9 +1,7 @@
 import { useMemo } from 'react'
 import { tv } from 'tailwind-variants'
 
-import { useEnvironment } from '../useEnvironment'
-
-import type { Environment } from '../useEnvironment'
+import { type Environment, useEnvironment } from '../client/useEnvironment'
 
 // tailwind-variantsのslotsから型を抽出する型ヘルパー
 type Slots<T> = T extends { slots: infer S }
